Updated UserContext for Home
Reorganized the app so that the App class is the root and contains the
UserContext context, which now stores user information and could also
store permission/class information for the user across the app. The
HomeController and subcomponents of Home now read the user from the
UserContext.  If the user is not set, then the HomeController itself
will display only the login page.

Note: needs to be applied to the other components as well.