Bootstrap and Reorganization
This commit updates much of the site to a new context-based storage for
the current user and the user's classes.  The home section works, select
course, and join queue page.  There is bare-minimum functionality in the
admin section.

The queue **DOES NOT** work yet, but this commit does allow work to
continue in other sections.